fixed the bug in the page of contributors in the sponsors
